qt5: add --qtlibsfromsys option to generate the list of Qt5 libraries by searching on the system
This option gives the possibility to search on the system (QT5_LIBDIR) for libraries available
instead of using the list hardcoded in the tool. The search is done using a regexp that matches
the same files as the ones used for library search with support for dynamic/static and win32/unix.

This option can be useful in a couple of situations:
-) A Qt5 version that is not (yet) supported by waf and therefore waf doesn't have the libraries yet
-) A custom Qt5 version with many unofficial libraries (that otherwise have to be added by hand
or using --qtextralibs)
-) A very small Qt5 installation (ie. embedded systems) where otherwise waf configure will anyway check
for the big list of libraries (and not find most of them)